fix relative pronoun
The last subphrase of this sentence is a relative clause.
Esperanto uses the question-correlatives (from the
tabelvortoj; see https://lernu.net/gramatiko/tabelvortoj)
as relative pronouns.
Thus "that" can only be translated as "ke" when
introducing a subphrase about assertions
(«She says that …» / «I believe that …»)
and a k-tabelvorto (here "kiu" or "kio") has to be used instead
as the translation of "ke" for introducing a relative clause.

Note that any relative pronoun here
(no matter whether one chooses "kio" or "kiu")
is ambiguous of whether it refers to just "la monato"
or (as intended here) to "la unua tago de la monato".
This could only be remedied by more profound rephrasing.